test/topology: split raft upgrade tests
Split raft upgrade tests to run in parallel by default
